# backend/docs/swagger-output.json
{
  "openapi": "2.0",
  "info": {
    "title": "GLY VTU API",
    "description": "Auto-generated Swagger specification for GLY VTU backend.",
    "version": "3.4.4"
  },
  "servers": [
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/{id}/messages": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/{id}/send": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Use: `Bearer <token>`"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}